--- a/include/xtensor/xstrided_view.hpp
+++ b/include/xtensor/xstrided_view.hpp
@@ -153,10 +153,36 @@ namespace xt
 
         using inner_storage_type = typename base_type::inner_storage_type;
         using storage_type = typename base_type::storage_type;
-        using linear_iterator = typename storage_type::iterator;
-        using const_linear_iterator = typename storage_type::const_iterator;
-        using reverse_linear_iterator = std::reverse_iterator<linear_iterator>;
-        using const_reverse_linear_iterator = std::reverse_iterator<const_linear_iterator>;
+
+        template <class C, class = void_t<>>
+        struct get_linear_iterator : std::false_type
+        {
+            using iterator = typename C::iterator;
+        };
+
+        template <typename C>
+        struct get_linear_iterator<C, void_t<decltype(std::declval<C>().linear_begin())>> : std::true_type
+        {
+            using iterator = typename C::linear_iterator;
+        };
+
+        template <class C, class = void_t<>>
+        struct get_const_linear_iterator : std::false_type
+        {
+            using iterator = typename C::const_iterator;
+        };
+
+        template <typename C>
+        struct get_const_linear_iterator<C, void_t<decltype(std::declval<C>().linear_cbegin())>> : std::true_type
+        {
+            using iterator = typename C::const_linear_iterator;
+        };
+
+        using linear_iterator = typename get_linear_iterator<storage_type>::iterator;
+        using const_linear_iterator = typename get_const_linear_iterator<storage_type>::iterator;
+        using reverse_linear_iterator = std::reverse_iterator<typename get_linear_iterator<storage_type>::iterator>;
+        using const_reverse_linear_iterator = std::reverse_iterator<
+            typename get_const_linear_iterator<storage_type>::iterator>;
 
         using iterable_base = select_iterable_base_t<L, xexpression_type::static_layout, self_type>;
         using inner_shape_type = typename base_type::inner_shape_type;
@@ -222,6 +248,7 @@ namespace xt
         const_linear_iterator linear_begin() const;
         const_linear_iterator linear_end() const;
         const_linear_iterator linear_cbegin() const;
+
         const_linear_iterator linear_cend() const;
 
         reverse_linear_iterator linear_rbegin();
@@ -511,7 +538,16 @@ namespace xt
     template <class CT, class S, layout_type L, class FST>
     inline auto xstrided_view<CT, S, L, FST>::linear_cbegin() const -> const_linear_iterator
     {
-        return this->storage().cbegin() + static_cast<std::ptrdiff_t>(data_offset());
+        return xtl::mpl::static_if<get_const_linear_iterator<storage_type>::value>(
+            [&](auto self)
+            {
+                return self(this->storage()).linear_cbegin() + static_cast<std::ptrdiff_t>(data_offset());
+            },
+            [&](auto self)
+            {
+                return self(this->storage()).cbegin() + static_cast<std::ptrdiff_t>(data_offset());
+            }
+        );
     }
 
     template <class CT, class S, layout_type L, class FST>