@spectre-ns 2024-07-11
@JohanMabille I'm trying to forward linear iterators where possible but I think the implementations expect steppers in certain places and there are some subtle differences I haven't figured out yet. I thought it was weird that the linear methods on the strided view actually return xiterators of steppers rather than linear iterators as the interface would suggest.

Test bench from issue #2734 

Without Changes:
case 1:  10.6235ms
case 2a: 48.4154ms
case 2b: 8.93694ms
case 2c: 48.7691ms
case 3a: 47.3316ms
case 3b: 9.05212ms
case 3c: 47.3312ms

With changes results in:
case 1:  10.0266ms
case 2a: 9.23594ms
case 2b: 8.97092ms
case 2c: 9.09802ms
case 3a: 8.86476ms
case 3b: 9.55695ms
case 3c: 8.67326ms

The implementation is super rough but I wanted to show you the basic idea. I ran into a pesky issue with reshaped views that resulted in my having to force a fallback to the old implementation. Might need some help there but the performance improvements speak for themselves.