@spectre-ns 2024-07-12
I fixed the issues with the reshape views. We cannot use linear iteration when we have dynamic layouts (Obviously). Needed to add a check in the reshape view constructor to get around this. Super ugly and I'm sure there is a better way.