@spectre-ns 2024-07-12
Can a linear assignment be used when the inner stride is not one? It seems like the linear iterators increment by one but when doing conversion from row major to column major they strides between adjacent elements is non-unity.