@spectre-ns 2024-07-19
> Thanks a lot for tackling this! I think the linear / not linear iteration was supposed to be handled in the flat_storage_adaptor but we probably missed some cases.
> 
> I wonder if it would make sense to merge the new `linear_flat_expression_adaptor` with the existing `flat_expression_adaptor`. This would put the reponsibility of choosing between the linear iterator and the regular one in the flat adaptor rather than in the strided_view. Also It would probably benefit other classes using the flat_expression adaptor.

@JohanMabille  I debated that as well. Let me play around with the logic and see if I can make this work without modifying the `strided_view` class and instead us the `flat_expression_adaptor` implementation. That would likely be cleaner. I was hoping to demonstrate the type of performance improvements before coming up with the final solution.