@spectre-ns 2024-07-21
> Thanks a lot for tackling this! I think the linear / not linear iteration was supposed to be handled in the flat_storage_adaptor but we probably missed some cases.
> 
> I wonder if it would make sense to merge the new `linear_flat_expression_adaptor` with the existing `flat_expression_adaptor`. This would put the reponsibility of choosing between the linear iterator and the regular one in the flat adaptor rather than in the strided_view. Also It would probably benefit other classes using the flat_expression adaptor.

@JohanMabille  combining them became difficult because the uvector implementation doesn't have the linear_iterator concept. The begin and end methods are linear iterators but the interface is not consistent. I'm not sure if this is intentional or not.